pyproject.toml
[project]
name = "analytics-narou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
